pyproject.toml
[tool.poetry]
name = "goroshek"
version = "0.1.0"
description = "Info channel management tool"
authors = ["Maxim Vasilev <vsmaxim8@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
natasha = "^1.3.0"
python-telegram-bot = "^12.8"
python-dotenv = "^0.14.0"
google-api-python-client = "^1.11.0"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
